+++ web_src/src/pages/workflowv2/dashboard/numberPanelFormConstants.ts
@@ -1,0 +1,13 @@
+import type { WidgetColumnFormat, WidgetNumberAggregation } from "./widget/types";
+
+export const NUMBER_PANEL_AGGREGATIONS: WidgetNumberAggregation[] = [
+  "count",
+  "sum",
+  "avg",
+  "min",
+  "max",
+  "first",
+  "last",
+];
+
+export const NUMBER_PANEL_FORMATS: WidgetColumnFormat[] = ["text", "number", "percent", "duration"];
